Create endpoint for eviction notices
This adapter uses the [Eviction
Notices](https://data.sfgov.org/Housing-and-Buildings/Eviction-Notices/5cei-gny5)
dataset from data.sfgov.org.

It works just like the rest of the socrata adapters in this repo. The
only notable difference is that I made a mapping of reasons for
evictions that turn them into more human text (I'm not a eviction
expert, so I copied most of the text from the tooltips on the socrata
page for this data).

Tests included.